# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GigWork — Gig-based local work marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /v1/users/me/role:
    put:
      operationId: switchRole
      tags: [users]
      summary: Switch between seeker and provider role
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "200":
          description: Role switched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /v1/users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get public profile of a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found

  # ── CATEGORIES ────────────────────────────────────────────────────────────────
...
  /v1/stats/provider-dashboard:
    get:
      operationId: getProviderDashboard
      tags: [stats]
      summary: Provider dashboard stats (active jobs, total hires, applications)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Provider dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderDashboardStats"
  /v1/stats/seeker-dashboard:
    get:
      operationId: getSeekerDashboard
      tags: [stats]
      summary: Seeker dashboard stats (applications, completions, rating)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Seeker dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeekerDashboardStats"
  /v1/stats/job-activity/{jobId}:
    get:
      operationId: getJobActivity
      tags: [stats]
      summary: Job activity breakdown (views, applications by day)
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job activity stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobActivity"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      required: [id, phone, current_role, created_at]
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: ["string", "null"]
        profile_photo_url:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        current_role:
          type: string
          enum: [seeker, provider]
        is_phone_verified:
          type: boolean
        is_id_verified:
          type: boolean
        trust_score:
          type: number
        is_banned:
          type: boolean
        seeker_profile:
          $ref: "#/components/schemas/SeekerProfile"
        provider_profile:
          $ref: "#/components/schemas/ProviderProfile"
        created_at:
          type: string
    SeekerProfile:
      type: object
      properties:
        bio:
          type: ["string", "null"]
        availability:
          type: string
          enum: [immediate, within_week, flexible]
        preferred_work_type:
          type: string
          enum: [hourly, daily, monthly, any]
        salary_expectation:
          type: ["integer", "null"]
        years_experience:
          type: integer
        total_jobs_completed:
          type: integer
        avg_rating:
          type: number
        total_ratings:
          type: integer
        attendance_score:
          type: number
        skills:
          type: array
          items:
            type: string
    ProviderProfile:
      type: object
      properties:
        business_name:
          type: ["string", "null"]
        business_type:
          type: ["string", "null"]
        total_jobs_posted:
          type: integer
        total_hires:
          type: integer
        avg_rating:
          type: number
        total_ratings:
          type: integer
        avg_response_hours:
          type: ["number", "null"]
        hiring_success_rate:
          type: number
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        profile_photo_url:
          type: string
        bio:
          type: string
        availability:
          type: string
          enum: [immediate, within_week, flexible]
        preferred_work_type:
          type: string
          enum: [hourly, daily, monthly, any]
        salary_expectation:
          type: integer
        years_experience:
          type: integer
        skills:
          type: array
          items:
            type: string
        business_name:
          type: string
        business_type:
          type: string
    RoleInput:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [seeker, provider]

    # ── CATEGORY SCHEMAS ─────────────────────────────────────────────────────────
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        name_hi:
          type: ["string", "null"]
        icon_name:
          type: ["string", "null"]
        sort_order:
          type: integer

    # ── JOB SCHEMAS ──────────────────────────────────────────────────────────────
    Job:
      type: object
      required: [id, title, work_type, city, status, created_at]
      properties:
        id:
          type: string
        provider_id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        category_id:
          type: ["integer", "null"]
        category:
          $ref: "#/components/schemas/Category"
        work_type:
          type: string
          enum: [hourly, daily, monthly]
        # hourly
        hourly_rate:
          type: ["integer", "null"]
        total_hours:
          type: ["integer", "null"]
        shift_start_time:
          type: ["string", "null"]
        shift_end_time:
          type: ["string", "null"]
        same_day_payment:
          type: ["boolean", "null"]
        urgent_hiring:
          type: ["boolean", "null"]
        # daily
        daily_wage:
          type: ["integer", "null"]
        num_working_days:
          type: ["integer", "null"]
        food_included:
          type: ["boolean", "null"]
        accommodation_included:
          type: ["boolean", "null"]
        overtime_available:
          type: ["boolean", "null"]
        # monthly
        monthly_salary:
          type: ["integer", "null"]
        joining_date:
          type: ["string", "null"]
        experience_required:
          type: ["integer", "null"]
        working_days_per_week:
          type: ["integer", "null"]
        pf_esi_included:
          type: ["boolean", "null"]
        salary_negotiable:
          type: ["boolean", "null"]
        # location
        city:
          type: string
        address:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        # status
        status:
          type: string
          enum: [draft, active, paused, filled, expired, removed]
        is_urgent:
          type: boolean
        is_verified:
          type: boolean
        views_count:
          type: integer
        applications_count:
          type: integer
        skills_required:
          type: array
          items:
            type: string
        provider:
          $ref: "#/components/schemas/UserProfile"
        expires_at:
          type: ["string", "null"]
        created_at:
          type: string
        has_applied:
          type: ["boolean", "null"]
...
    Application:
      type: object
      required: [id, job_id, seeker_id, status, applied_at]
      properties:
        id:
          type: string
        job_id:
          type: string
        seeker_id:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, withdrawn, completed]
        employer_note:
          type: ["string", "null"]
        applied_at:
          type: string
        responded_at:
          type: ["string", "null"]
        completed_at:
          type: ["string", "null"]
        job:
          $ref: "#/components/schemas/Job"
        seeker:
          $ref: "#/components/schemas/UserProfile"
        chat_room_id:
          type: ["string", "null"]
...
    ProviderDashboardStats:
      type: object
      required: [active_jobs, total_applications, total_hires, pending_applications, jobs_by_status]
      properties:
        active_jobs:
          type: integer
        total_applications:
          type: integer
        total_hires:
          type: integer
        pending_applications:
          type: integer
        avg_rating:
          type: number
        hiring_success_rate:
          type: number
        jobs_by_status:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
        recent_applications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
    SeekerDashboardStats:
      type: object
      required: [total_applied, pending_count, accepted_count, completed_count, avg_rating]
      properties:
        total_applied:
          type: integer
        pending_count:
          type: integer
        accepted_count:
          type: integer
        completed_count:
          type: integer
        avg_rating:
          type: number
        trust_score:
          type: number
        total_jobs_completed:
          type: integer
        attendance_score:
          type: number
    StatusCount:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer
    JobActivity:
      type: object
      required: [job_id, total_views, total_applications, applications_by_day]
      properties:
        job_id:
          type: string
        total_views:
          type: integer
        total_applications:
          type: integer
        applications_by_day:
          type: array
          items:
            $ref: "#/components/schemas/DayCount"
        status_breakdown:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
    DayCount:
      type: object
      required: [date, count]
      properties:
        date:
          type: string
        count:
          type: integer