# lib/api-zod/src/generated/types/providerDashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GigWork — Gig-based local work marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from './application';
import type { StatusCount } from './statusCount';

export interface ProviderDashboardStats {
  active_jobs: number;
  total_applications: number;
  total_hires: number;
  pending_applications: number;
  avg_rating?: number;
  hiring_success_rate?: number;
  jobs_by_status: StatusCount[];
  recent_applications?: Application[];
}