# lib/api-zod/src/generated/types/seekerDashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GigWork — Gig-based local work marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface SeekerDashboardStats {
  total_applied: number;
  pending_count: number;
  accepted_count: number;
  completed_count: number;
  avg_rating: number;
  trust_score?: number;
  total_jobs_completed?: number;
  attendance_score?: number;
}